delete and put

+++ src/main/java/com/example/app/services/ProductServiceImplementation.java
@@ -2,6 +2,7 @@ package com.example.app.services;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale.Category;
 import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
@@ -129,20 +130,28 @@ public class ProductServiceImplementation implements IProductService {
     }
 
     @Override
-    @Transactional(readOnly = true)
+    @Transactional
     public ResponseEntity<productResponseRest> deleteById(Long id) {
         productResponseRest response = new productResponseRest();
 
         try {
+            Optional<Product> productSearch = productDao.findById(id);
 
-            this.productDao.deleteById(id);
-            response.setMetadata("Respuesta OK", "200", "Producto eliminado");
-            return new ResponseEntity<>(response, HttpStatus.OK);
+            if (productSearch.isPresent()) {
+                this.productDao.deleteById(id);
+                response.setMetadata("Respuesta OK", "200", "Producto eliminado");
+                return new ResponseEntity<>(response, HttpStatus.OK);
+
+            } else {
+                response.setMetadata("Respuesta ERROR", "404", "Producto no encontrado");
+                return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+            }
 
         } catch (Exception e) {
             response.setMetadata("Respuesta ERROR", "500", "Error al eliminar producto");
             return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
         }
+
     }
 
     @Override
@@ -170,54 +179,117 @@ public class ProductServiceImplementation implements IProductService {
 
     }
 
+    // @Override
+    // @Transactional
+    // public ResponseEntity<productResponseRest> updateProduct(Product product,
+    // Long id) {
+    // productResponseRest response = new productResponseRest();
+    // List<Product> list = new ArrayList<>();
+
+    // // try {
+    // Optional<Categoria> categoria =
+    // this.categoriaDao.findById(product.getCategoria().getId());
+
+    // if (categoria.isPresent()) {
+    // product.setCategoria(categoria.get());
+    // } else {
+
+    // response.setMetadata("Respuesta ERROR", "404", "Categoria no encontrada");
+    // return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+    // }
+
+    // Optional<Product> productSearch = this.productDao.findById(id);
+
+    // if (productSearch.isPresent()) {
+
+    // productSearch.get().setName(product.getName());
+    // productSearch.get().setPrice(product.getPrice());
+    // productSearch.get().setCantidad(product.getCantidad());
+    // productSearch.get().setCategoria(product.getCategoria());
+    // productSearch.get().setPhoto(product.getPhoto());
+    // Product productUpdate = this.productDao.save(productSearch.get());
+    // if (productUpdate != null) {
+    // list.add(productUpdate);
+    // response.getProductResponse().setProducts(list);
+    // response.setMetadata("Respuesta OK", "200", "Producto actualizado con
+    // exito");
+
+    // } else {
+    // response.setMetadata("Respuesta ERROR", "400", "Error al actualizar
+    // producto");
+    // return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+    // }
+
+    // } else {
+    // response.setMetadata("Respuesta ERROR", "404", "Producto no encontrado");
+    // return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+    // }
+
+    // // } catch (Exception e) {
+    // // response.setMetadata("Respuesta ERROR", "500", "Error 500");
+    // // return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+    // // }
+    // return new ResponseEntity<>(response, HttpStatus.OK);
+
+    // }
+
     @Override
     @Transactional
-    public ResponseEntity<productResponseRest> updateProduct(Product product, Long id) {
+    public ResponseEntity<productResponseRest> updateProduct(Product product, Long categoryId, Long id) {
         productResponseRest response = new productResponseRest();
         List<Product> list = new ArrayList<>();
 
         try {
-            Optional<Categoria> categoria = this.categoriaDao.findById(product.getCategoria().getId());
 
-            if (categoria.isPresent()) {
-                product.setCategoria(categoria.get());
+            // search category to set in the product object
+            Optional<Categoria> category = this.categoriaDao.findById(categoryId);
+
+            if (category.isPresent()) {
+                product.setCategoria(category.get());
             } else {
-
-                response.setMetadata("Respuesta ERROR", "404", "Categoria no encontrada");
-                return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
-            }
-
-            Optional<Product> productSearch = this.productDao.findById(id);
-
-            if (productSearch.isPresent()) {
-
-                productSearch.get().setName(product.getName());
-                productSearch.get().setPrice(product.getPrice());
-                productSearch.get().setCantidad(product.getCantidad());
-                productSearch.get().setCategoria(product.getCategoria());
-                productSearch.get().setPhoto(product.getPhoto());
-                Product productUpdate = this.productDao.save(productSearch.get());
-                if (productUpdate != null) {
-                    list.add(productUpdate);
-                    response.getProductResponse().setProducts(list);
-                    response.setMetadata("Respuesta OK", "200", "Producto actualizado con exito");
-
-                } else {
-                    response.setMetadata("Respuesta ERROR", "400", "Error al actualizar producto");
-                    return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
-                }
-
-            } else {
-                response.setMetadata("Respuesta ERROR", "404", "Producto no encontrado");
+                response.setMetadata("respuesta nok", "-1", "Categoria no encontrada asociada al producto ");
                 return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
             }
 
-        } catch (Exception e) {
-            response.setMetadata("Respuesta ERROR", "500", "Error 500");
-            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-        return new ResponseEntity<>(response, HttpStatus.OK);
+            // search Product to update
+            Optional<Product> productSearch = productDao.findById(id);
 
+            if (productSearch.isPresent()) {
+
+                // se actualizará el producto
+                productSearch.get().setCantidad(product.getCantidad());
+                productSearch.get().setCategoria(product.getCategoria());
+                productSearch.get().setName(product.getName());
+                productSearch.get().setPhoto(product.getPhoto());
+                productSearch.get().setPrice(product.getPrice());
+
+                // save the product in DB
+                Product productToUpdate = productDao.save(productSearch.get());
+
+                if (productToUpdate != null) {
+                    list.add(productToUpdate);
+                    response.getProductResponse().setProducts(list);
+                    response.setMetadata("respuesta ok", "00", "Producto actualizado");
+                } else {
+                    response.setMetadata("respuesta nok", "-1", "Producto no actualizado ");
+                    return new ResponseEntity<>(response, HttpStatus.BAD_REQUEST);
+
+                }
+
+            } else {
+                response.setMetadata("respuesta nok", "-1", "Producto no actualizado ");
+                return new ResponseEntity<>(response, HttpStatus.NOT_FOUND);
+
+            }
+
+        } catch (Exception e) {
+            e.getStackTrace();
+            response.setMetadata("respuesta nok", "-1", "Error al actualizar producto");
+            return new ResponseEntity<>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+
+        }
+
+        return new ResponseEntity<>(response, HttpStatus.OK);
     }
 
 }
